requirements.txt
Django~=5.1.3
django-cleanup~=9.0.0
djangorestframework~=3.15.2
djangorestframework_simplejwt~=5.4.0
django-jazzmin~=3.0.1
django-filter~=24.3
drf-yasg~=1.21.8
psycopg2-binary~=2.9.10